mt_llm_ctx/get_ctx_params(): Add explicit cast to avoid compiler warning.

+++ mt_llm/mt_llm_ctx.cpp
@@ -56,7 +56,7 @@ static llama_context_params get_ctx_params(
         }
         else
         {
-            if(mt_p.n_ctx < model_n_ctx_train)
+            if(mt_p.n_ctx < static_cast<uint32_t>(model_n_ctx_train))
             {
                 // Use model's context length (it is the minimum).
